+++ src/main/java/bowling/step2/domain/pitchresult/factory/PitchResultGroupFactory.java
@@ -1,10 +1,10 @@
-package bowling.step2.outputview.pitchresult.factory;
+package bowling.step2.domain.pitchresult.factory;
 
 import bowling.step2.domain.Frame;
 import bowling.step2.domain.NormalFrame;
-import bowling.step2.outputview.pitchresult.LastFramePitchResultGroup;
-import bowling.step2.outputview.pitchresult.NormalFramePitchResultGroup;
-import bowling.step2.outputview.pitchresult.PitchResultGroup;
+import bowling.step2.domain.pitchresult.LastFramePitchResultGroup;
+import bowling.step2.domain.pitchresult.NormalFramePitchResultGroup;
+import bowling.step2.domain.pitchresult.PitchResultGroup;
 
 public class PitchResultGroupFactory {
     public static PitchResultGroup create(Frame frame) {
