+++ src/main/java/bowling/step2/domain/state/Gutter.java
@@ -1,4 +1,4 @@
-package bowling.step2.outputview.state;
+package bowling.step2.domain.state;
 
 public class Gutter implements PitchStatus{
     @Override
