+++ src/main/java/bowling/step2/outputview/OutputView.java
@@ -2,7 +2,7 @@ package bowling.step2.outputview;
 
 import bowling.step2.domain.Frame;
 import bowling.step2.domain.Lane;
-import bowling.step2.outputview.pitchresult.FrameResult;
+import bowling.step2.domain.pitchresult.FrameResult;
 
 import java.util.List;
